pd: rename Overlay extension traits to View
