+++ pd/src/components/app.rs
@@ -1,12 +1,18 @@
-use anyhow::Result;
+use std::str::FromStr;
+
+use anyhow::{anyhow, Result};
 use async_trait::async_trait;
 use jmt::{RootHash, Version};
+use penumbra_chain::params::ChainParams;
+use penumbra_stake::Epoch;
 use penumbra_transaction::Transaction;
 use tendermint::abci::{self, types::ValidatorUpdate};
+use tendermint::Time;
 use tracing::instrument;
 
+use crate::{genesis, Storage, WriteOverlayExt};
+
 use super::{Component, IBCComponent, Overlay, ShieldedPool, Staking};
-use crate::{genesis, PenumbraStore, Storage, WriteOverlayExt};
 
 /// The Penumbra application, written as a bundle of [`Component`]s.
 ///
@@ -127,3 +133,75 @@ impl Component for App {
         Ok(())
     }
 }
+
+/// This trait provides read and write access to common parts of the Penumbra
+/// state store.
+///
+/// Note: the `get_` methods in this trait assume that the state store has been
+/// initialized, so they will error on an empty state.
+#[async_trait]
+pub trait View: WriteOverlayExt {
+    /// Gets the chain parameters from the JMT.
+    async fn get_chain_params(&self) -> Result<ChainParams> {
+        self.get_domain(b"chain_params".into())
+            .await?
+            .ok_or_else(|| anyhow!("Missing ChainParams"))
+    }
+
+    /// Writes the provided chain parameters to the JMT.
+    async fn put_chain_params(&self, params: ChainParams) {
+        self.put_domain(b"chain_params".into(), params).await
+    }
+
+    /// Gets the current epoch for the chain.
+    async fn get_current_epoch(&self) -> Result<Epoch> {
+        let block_height = self.get_block_height().await?;
+        Ok(Epoch::from_height(
+            block_height,
+            self.get_epoch_duration().await?,
+        ))
+    }
+
+    /// Gets the epoch duration for the chain.
+    async fn get_epoch_duration(&self) -> Result<u64> {
+        // this might be a bit wasteful -- does it matter?  who knows, at this
+        // point. but having it be a separate method means we can do a narrower
+        // load later if we want
+        self.get_chain_params()
+            .await
+            .map(|params| params.epoch_duration)
+    }
+
+    /// Gets the current block height from the JMT
+    async fn get_block_height(&self) -> Result<u64> {
+        let height_bytes: u64 = self
+            .get_proto(b"block_height".into())
+            .await?
+            .ok_or_else(|| anyhow!("Missing block_height"))?;
+
+        Ok(height_bytes)
+    }
+
+    /// Writes the block height to the JMT
+    async fn put_block_height(&self, height: u64) {
+        self.put_proto(b"block_height".into(), height).await
+    }
+
+    /// Gets the current block timestamp from the JMT
+    async fn get_block_timestamp(&self) -> Result<Time> {
+        let timestamp_string: String = self
+            .get_proto(b"block_timestamp".into())
+            .await?
+            .ok_or_else(|| anyhow!("Missing block_timestamp"))?;
+
+        Ok(Time::from_str(&timestamp_string).unwrap())
+    }
+
+    /// Writes the block timestamp to the JMT
+    async fn put_block_timestamp(&self, timestamp: Time) {
+        self.put_proto(b"block_timestamp".into(), timestamp.to_rfc3339())
+            .await
+    }
+}
+
+impl<T: WriteOverlayExt> View for T {}
